feat: add validation results for GetContactPolicies endpoint

+++ Doppler.ContactPolicies.Data.Access/Entities/User.cs
@@ -8,7 +8,7 @@ namespace Doppler.ContactPolicies.Data.Access.Entities
 {
     public class User
     {
-        public long IdUser { get; set; }
+        public int IdUser { get; set; }
         public string Email { get; set; }
     }
 }
